+++ src/components/menu/component.test-pw.tsx
@@ -250,7 +250,7 @@ export const MenuComponentFullScreen = (
 export const MenuFullScreenBackgroundScrollTest = () => {
   return (
     <Box height="2000px" position="relative">
-      <Box height="100px" id="bottom-box" position="absolute" bottom="0px">
+      <Box height="100px" position="absolute" bottom="0px">
         I should not be scrolled into view
       </Box>
       <MenuFullscreen isOpen onClose={() => {}}>
@@ -482,6 +482,41 @@ export const MenuSegmentTitleComponent = (props: Partial<MenuTitleProps>) => {
               <MenuItem onClick={() => {}}>Item Submenu One</MenuItem>
               <MenuSegmentTitle {...props} text="segment title" />
               <MenuItem href="#">Item Submenu Two</MenuItem>
+            </MenuItem>
+          </Menu>
+        </div>
+      ))}
+    </Box>
+  );
+};
+
+export const MenuSegmentTitleComponentWithAdditionalMenuItem = (
+  props: Partial<MenuTitleProps>
+) => {
+  return (
+    <Box mb={150}>
+      {menuTypes.map((menuType) => (
+        <div key={menuType}>
+          <Typography variant="h4" textTransform="capitalize" my={2}>
+            {menuType}
+          </Typography>
+          <Menu menuType={menuType} display="flex">
+            <MenuItem href="#">Menu Item One</MenuItem>
+            <MenuItem href="#">Menu Item Two</MenuItem>
+            <MenuItem submenu="Menu Item Three">
+              <MenuItem href="#">Item Submenu One</MenuItem>
+              <MenuItem href="#">Item Submenu Two</MenuItem>
+              <MenuItem icon="settings" href="#">
+                Item Submenu Three
+              </MenuItem>
+              <MenuItem href="#">Item Submenu Four</MenuItem>
+            </MenuItem>
+            <MenuItem submenu="Menu Item Four" onClick={() => {}}>
+              <MenuItem onClick={() => {}}>Item Submenu One</MenuItem>
+              <MenuSegmentTitle {...props} text="segment title">
+                <MenuItem href="#">Last Segment Child</MenuItem>
+              </MenuSegmentTitle>
+              <MenuItem href="#">Menu Item Five</MenuItem>
             </MenuItem>
           </Menu>
         </div>