+++ src/components/select/filterable-select/filterable-select.spec.tsx
@@ -278,35 +278,6 @@ describe("FilterableSelect", () => {
     });
   });
 
-  describe("when the inputRef function prop is specified", () => {
-    it("should display deprecation warning when the inputRef prop is used", () => {
-      const inputRefFn = jest.fn();
-      const wrapper = renderSelect({ inputRef: inputRefFn });
-
-      expect(loggerSpy.mock.calls).toEqual([
-        [
-          "The `inputRef` prop in `Filterable Select` component is deprecated and will soon be removed. Please use `ref` instead.",
-        ],
-        [
-          "The `inputRef` prop in `Textbox` component is deprecated and will soon be removed. Please use `ref` instead.",
-        ],
-      ]);
-      expect(loggerSpy).toHaveBeenCalledTimes(2);
-      // will be called twice because the prop is passed to Textbox where another deprecation warning is triggered.
-      wrapper.setProps({ prop1: true });
-      expect(loggerSpy).toHaveBeenCalledTimes(2);
-    });
-
-    it("then the input reference should be returned on call", () => {
-      const inputRefFn = jest.fn();
-      const wrapper = renderSelect({ inputRef: inputRefFn });
-
-      expect(inputRefFn).toHaveBeenCalledWith({
-        current: wrapper.find("input").getDOMNode(),
-      });
-    });
-  });
-
   describe("when the onFocus prop has been passed and the input has been focused", () => {
     it("then that prop should be called", () => {
       const onFocusFn = jest.fn();