+++ src/components/switch/switch.spec.tsx
@@ -109,19 +109,6 @@ describe("Switch", () => {
   describe("refs", () => {
     let wrapper: ReactWrapper;
 
-    it("should display deprecation warning when the inputRef prop is used", () => {
-      const ref = { current: null };
-
-      wrapper = render({ inputRef: ref });
-
-      expect(loggerSpy).toHaveBeenCalledWith(
-        "The `inputRef` prop in `Switch` component is deprecated and will soon be removed. Please use `ref` instead."
-      );
-      wrapper.setProps({ prop1: true });
-      expect(loggerSpy).toHaveBeenCalledTimes(1);
-      loggerSpy.mockRestore();
-    });
-
     it("accepts ref as a ref object", () => {
       const ref = { current: null };
       wrapper = render({ ref });